requirements.txt
GitPython==3.1.45
requests==2.32.5
